+++ Forge/changelog.md
@@ -1,3 +1,6 @@
-1.4.2:
+1.5.0:
+* Added White Fabric
+* Added recipe for Seed Satchel
+* Added Tomato Splat sound
 * Tweaked Pickle recipe
 * Added recipe to Sugar Beets